requirements.txt
boto3==1.42.72
botocore==1.42.72
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
jmespath==1.1.0
kafka-python==2.3.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
setuptools==68.0.0
six==1.17.0
urllib3==2.6.3
